refactor(miniapp): remove utils file
