+++ packages/build-plugin-rax-app/src/plugins/MiniAppPlugin/tool/includes.js
@@ -1,0 +1,12 @@
+/**
+ * Judge whether parentArr includes childArr
+ */
+function includes(parentArr, childArr) {
+  for (const child of childArr) {
+    if (parentArr.indexOf(child) === -1) return false;
+  }
+
+  return true;
+}
+
+module.exports = includes;
